requirements.txt
typing
mysql-connector-python
python-dotenv
tabulate

